requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multi-key-dict==2.0.3
packaging==24.2
pbr==6.1.1
python-dateutil==2.9.0.post0
python-jenkins==1.8.2
pytz==2025.1
requests==2.32.3
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
